+++ src/main/java/net/sf/mpxj/LocaleData.java
@@ -2518,6 +2518,7 @@ public final class LocaleData extends ListResourceBundle
       PROJECT_COLUMNS_ARRAY[ProjectField.FILE_APPLICATION.getValue()] = "File Application";
       PROJECT_COLUMNS_ARRAY[ProjectField.FILE_TYPE.getValue()] = "File Type";
       PROJECT_COLUMNS_ARRAY[ProjectField.EXPORT_FLAG.getValue()] = "Export Flag";
+      PROJECT_COLUMNS_ARRAY[ProjectField.GUID.getValue()] = "GUID";
 
       PROJECT_COLUMNS_ARRAY[ProjectField.ENTERPRISE_CUSTOM_FIELD1.getValue()] = "Enterprise Custom Field 1";
       PROJECT_COLUMNS_ARRAY[ProjectField.ENTERPRISE_CUSTOM_FIELD2.getValue()] = "Enterprise Custom Field 2";