+++ src/main/java/net/sf/mpxj/primavera/PrimaveraDatabaseReader.java
@@ -197,7 +197,7 @@ public final class PrimaveraDatabaseReader extends AbstractProjectReader
       // Process common attributes
       //
       List<Row> rows = getRows("select * from " + m_schema + "project where proj_id=?", m_projectID);
-      m_reader.processProjectProperties(rows, m_projectID);
+      m_reader.processProjectProperties(rows);
 
       //
       // Process PMDB-specific attributes