+++ src/main/java/net/sf/mpxj/primavera/PrimaveraPMFileReader.java
@@ -221,8 +221,7 @@ public final class PrimaveraPMFileReader extends AbstractProjectStreamReader
          }
          else
          {
-            String uniqueID = m_projectID.toString();
-            project = projects.stream().filter(p -> uniqueID.equals(p.getProjectProperties().getUniqueID())).findFirst().orElse(null);
+            project = projects.stream().filter(p -> m_projectID.equals(p.getProjectProperties().getUniqueID())).findFirst().orElse(null);
 
          }
       }
@@ -640,11 +639,12 @@ public final class PrimaveraPMFileReader extends AbstractProjectStreamReader
 
       properties.setCreationDate(project.getCreateDate());
       properties.setFinishDate(project.getFinishDate());
+      properties.setGUID(DatatypeConverter.parseUUID(project.getGUID()));
       properties.setName(project.getName());
       properties.setStartDate(project.getPlannedStartDate());
       properties.setStatusDate(project.getDataDate());
       properties.setProjectTitle(project.getId());
-      properties.setUniqueID(project.getObjectId() == null ? null : project.getObjectId().toString());
+      properties.setUniqueID(project.getObjectId());
       properties.setExportFlag(!BooleanHelper.getBoolean(project.isExternal()));
       processScheduleOptions(project.getScheduleOptions());
    }
@@ -659,7 +659,7 @@ public final class PrimaveraPMFileReader extends AbstractProjectStreamReader
       properties.setStartDate(project.getPlannedStartDate());
       properties.setStatusDate(project.getDataDate());
       properties.setProjectTitle(project.getId());
-      properties.setUniqueID(project.getObjectId() == null ? null : project.getObjectId().toString());
+      properties.setUniqueID(project.getObjectId());
       properties.setExportFlag(false);
    }